+++ src/main/java/com/sangto/rental_car_server/domain/dto/booking/BookingResponseDTO.java
@@ -3,15 +3,17 @@ package com.sangto.rental_car_server.domain.dto.booking;
 import com.sangto.rental_car_server.domain.enums.EBookingStatus;
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record BookingResponseDTO(
         Integer bookingId,
         Integer carId,
         String carName,
-        Double basePrice,
+        BigDecimal basePrice,
         String startDateTime,
         String endDateTime,
-        Double totalPrice,
+        BigDecimal totalPrice,
         EBookingStatus bookingStatus
 ) {
 }
